feat: cached routes to support native currencies (#702)
- **What kind of change does this PR introduce?** (Bug fix, feature, docs update, ...)
feature

- **What is the current behavior?** (You can also link to an open issue here)
cached routes do not support native currency routing

- **What is the new behavior (if this is a feature change)?**
we want cached routes to support native currency routing. This is import for v4 native currency routing support.

- **Other information**:
Although we changed the method signature to accept Currency, for now we still pass in Tokens for tokenIn and tokenOut, so that this change should still remain no-op for now.