fix: ROUTE-489 Refresh pools in routes' metadata (#879)
* ROUTE-489 Refresh pools in routes' metadata

This PR refreshes the pools in the route object from swapRouteFromCache.
The offline lambda is never in the cache mode and the refresh helps to prevent stale data.

4.21.13

removed the mutable class, clone and replay the in-memory route object, adding metric, handle non-refreshed pools by pushing back to the queue

set retainFakePools to true

4.21.16

* rebasing

* formatting

* formatting